fix build error

+++ arch/h8300/kernel/setup.c
@@ -55,7 +55,7 @@ unsigned long memory_end;
 char __initdata command_line[COMMAND_LINE_SIZE];
 
 extern int _stext, _etext, _sdata, _edata, _sbss, _ebss, _end;
-extern int _ramstart, _ramend;
+extern unsigned long _ramstart, _ramend;
 extern char _target_name[];
 extern char command_start[];
 extern void h8300_gpio_init(void);
